Work a reading's projection out where it is asked for, and run its checks
A reading walked its term as it was made, to hold what that walk hashed to. What
a caller substitutes its arguments into is read as an assumption and then let go,
and nothing compares one, so every rule at every call site paid for a projection
nothing asks for. The term is the only thing held now, and what it says is worked
out where a reading is compared or hashed.

The checks that hold the projection asked it of fixtures and of two compilations
of every corpus, all under the tag the nightly runs. Each of the five things that
say where a node stands is held by a pair of fixtures, and a pair is what a
mutation of the projection has to be caught by — so those run with the build, and
what a corpus alone can answer, which kinds of term anything writes at all, stays
where a corpus is compiled twice.
