+++ souther-compiler/src/test/java/souther/compiler/check/EveryTermIsReadForWhatItSaysTest.java
@@ -2,12 +2,8 @@ package souther.compiler.check;
 
 import souther.compiler.DefaultStdlib;
 import souther.compiler.KeptCalls;
-import souther.compiler.conformance.ConformanceCorpus;
 import souther.compiler.core.Core;
 import souther.compiler.diag.SourcePos;
-import souther.compiler.meta.ModulePath;
-import souther.compiler.query.Bodies;
-import souther.compiler.query.Compilation;
 import souther.compiler.query.ReadAs;
 import souther.compiler.types.ApplicationOrigin;
 import souther.compiler.types.BinOp;
@@ -22,16 +18,9 @@ import souther.compiler.types.Type;
 import souther.compiler.types.ValueName;
 import souther.compiler.types.WrittenOwner;
 
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.TreeSet;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -51,196 +40,18 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * name and application a kept call was written as. Each is asked twice — moving it leaves the
  * reading alone, and moving what the node says does not.
  *
- * <p>Which kinds of term this reaches is measured rather than assumed. The corpus is a set of
- * models, and a model has no reason to write every kind, so a module written to reach the rest
- * stands beside it and what is still unreached is named.
+ * <p>Of fixtures, so that a build runs it. Every one of these is a pair of terms differing in one
+ * thing, which is what a mutation of the projection has to be caught by; asking it of a corpus
+ * instead would only reach the things two compilations of one model can be made to differ in, and
+ * moving every line of a file reaches the positions and no ordinal. What a corpus does answer —
+ * which kinds of term anything writes at all — is
+ * {@code EveryKindOfTermACorpusWritesIsReadForWhatItSaysTest}.
  */
-@Tag("population")
 class EveryTermIsReadForWhatItSaysTest {
 
     private static final SourcePos POS = new SourcePos(1, 1);
     private static final SourcePos ELSEWHERE = new SourcePos(9, 4);
     private static final WrittenOwner OWNER = new WrittenOwner.Body("demo", "b");
-
-    /**
-     * A module written to reach what the corpus does not. The corpus is a set of models, and a model
-     * has no reason to write every kind of term; this has no reason to be a model.
-     */
-    private static final String WIDE = """
-            module wide.terms exposing ( In, Out, Small, run )
-
-            import List ( all, get, find, length )
-            import Option ( map, withDefault )
-
-            data Small = Int
-                invariant value > 0
-
-            data In  = { xs: List<Int>, s: String, k: Int, note: String?, on: Date }
-            data Out = { n: Int, label: String, pair: Int, kept: String? }
-            data Bag = { items: List<Int> }
-
-            data Yes = { n: Int }
-            data No  = { n: Int }
-
-            let doubled (v) = v * 2
-
-
-            behavior bagged : (i: In) -> Bag
-                constructs Bag
-                ensures length(value.items) >= i.k - i.k
-            let bagged (i) = Bag { items = i.xs }
-
-            behavior pick : (i: In) -> Yes | No
-                constructs Yes, No
-            let pick (i) = if i.k > 0 then Yes { n = i.k } else No { n = 0 - i.k }
-
-            behavior only : (i: In) -> Small
-                constructs Small
-            let only (i) = match pick(i) with
-                | Yes -> Small(1)
-                | No  -> unreachable "the corpus only ever hands this a positive"
-
-            behavior run : (i: In) -> Out
-                constructs Out, Small
-            let run (i) = {
-                let negated = -doubled(i.k)
-                let some = get(0, i.xs) |> map(n -> n + 1) |> withDefault(0)
-                let none = find(n -> n > 1000000, i.xs) |> map(n -> n) |> withDefault(0)
-                let every = all(n -> n >= 0, i.xs)
-                let (left, right) = (some, none)
-                let noted = match i.note with
-                    | Some t -> String.length(t)
-                    | None   -> 0
-                let picked = if Small(negated) as ok then ok.value else noted
-                let fs: List<(Int) -> Int> = [(x) -> x + 1, (x) -> x + 2]
-                let applied = all((f) -> f(i.k) > 0, fs)
-                let dated = if i.on < Date("2026-01-01") then 1 else 0
-                Out {
-                    n = left + right + picked + length([1, 2, 3])
-                            + (if every then 1 else 0) + (if applied then 1 else 0) + dated,
-                    label = "fixed",
-                    kept = "here",
-                    pair = negated
-                }
-            }
-            """;
-
-    /** Every checked body of every corpus, and the same again with every line moved down. */
-    private static List<Core> bodies(String before) {
-        List<Core> out = new ArrayList<>();
-        List<Map<String, String>> workspaces = new ArrayList<>();
-        for (ConformanceCorpus corpus : ConformanceCorpus.all()) {
-            Map<String, String> byId = new LinkedHashMap<>();
-            for (int i = 0; i < corpus.sources().size(); i++) {
-                byId.put(corpus.files().get(i), before + corpus.sources().get(i));
-            }
-            workspaces.add(byId);
-        }
-        workspaces.add(Map.of("wide.sou", before + WIDE));
-        boolean wide = false;
-        for (Map<String, String> byId : workspaces) {
-            wide = byId.containsKey("wide.sou");
-            Compilation c = Compilation.ofDocuments(byId, Set.of(), ModulePath.EMPTY);
-            c.answerEverything();
-            if (wide) {
-                assertEquals(List.of(), c.db().allReports().stream().map(Object::toString).toList(),
-                        "the module written to reach the rest of the kinds compiles");
-            }
-            for (String module : c.modules()) {
-                // The behaviors a module declares, asked through the accessor a build asks it
-                // through. There was a query of its own for this and nothing but this reached it.
-                Set<String> names = c.declaredBehaviors(module);
-                if (names == null) {
-                    continue;
-                }
-                for (String behavior : new TreeSet<>(names)) {
-                    Bodies.CheckedBody checked =
-                            c.db().ask(new Bodies.CheckedBehavior(module, behavior)).value();
-                    if (checked != null && checked.body() != null) {
-                        out.add(checked.body());
-                    }
-                }
-                Map<String, StatedContract> stated =
-                        c.db().ask(new Bodies.StatedContracts(module)).value();
-                if (stated == null) {
-                    continue;
-                }
-                for (String behavior : new TreeSet<>(stated.keySet())) {
-                    for (StatedContract.StatedRule rule : stated.get(behavior).rules()) {
-                        for (StatedContract.Conjunct each : rule.conjuncts()) {
-                            if (each.stated().orNull() != null) {
-                                out.add(each.stated().orNull());
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        return out;
-    }
-
-    private static void each(Core e, java.util.function.Consumer<Core> f) {
-        if (e == null) {
-            return;
-        }
-        f.accept(e);
-        Core.forEachChild(e, child -> each(child, f));
-    }
-
-    private static Set<String> kindsIn(List<Core> terms) {
-        Set<Class<?>> found = new LinkedHashSet<>();
-        for (Core term : terms) {
-            each(term, node -> found.add(node.getClass()));
-        }
-        return named(found);
-    }
-
-    private static Set<String> named(Set<Class<?>> kinds) {
-        Set<String> out = new TreeSet<>();
-        for (Class<?> kind : kinds) {
-            out.add(kind.getSimpleName());
-        }
-        return out;
-    }
-
-    private static List<TermMeaning> read(List<Core> terms) {
-        return terms.stream().map(TermMeaning::of).toList();
-    }
-
-    /**
-     * The one kind no source reaches, and why.
-     *
-     * <p>A model cannot write an absence. An optional stands on a data field, a construction has to
-     * give that field a value, and there is no way to spell the empty one outside a fixture — which
-     * is never elaborated, so it makes no term. E1402 says as much where a model tries: answer a
-     * list of nought or one instead. So this is here rather than reached, and a kind that turns up
-     * beside it is one someone has to say the same about.
-     */
-    private static final Set<String> WRITTEN_BY_NOTHING = new TreeSet<>(Set.of("OptionNone"));
-
-    @Test
-    void everyKindOfTermTheCorpusWritesIsReached() {
-        Set<String> reached = kindsIn(bodies(""));
-        Set<String> declared = named(Set.of(Core.class.getPermittedSubclasses()));
-        declared.removeAll(reached);
-
-        assertEquals(WRITTEN_BY_NOTHING, declared,
-                "a kind of term nothing here reaches is one nothing here reads");
-    }
-
-    /**
-     * And two readings of one corpus that differ only in where its lines are come to one value. The
-     * property the rest of this is for: a caller depending on what a term says is not an edit away
-     * from a blank line somewhere above it.
-     */
-    @Test
-    void movingEveryLineChangesNoReading() {
-        List<TermMeaning> where = read(bodies(""));
-        List<TermMeaning> moved = read(bodies("\n\n\n"));
-
-        assertEquals(where.size(), moved.size(), "the same bodies compile either way");
-        assertEquals(where, moved, "and each says what it said, three lines further down");
-    }
 
     @Test
     void aPositionIsNotRead() {
